chore: deduplicate transformation print code
